Move more help logic into helptext.py; more functions into util.py
